backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
feedparser==6.0.10
requests==2.28.2
# Production dependencies
gunicorn==20.1.0
psycopg2-binary==2.9.5
Flask-SQLAlchemy==3.0.3
alembic==1.10.4

# Google AI dependencies
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
google-generativeai>=0.3.0

# !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
# !!! PYTHON 3.13+ COMPATIBILITY NOTICE                                          !!!
# !!! The following packages are not compatible with Python 3.13+                !!!
# !!! They have been commented out to allow basic functionality                  !!!
# !!! For full functionality, please use Python 3.10                             !!!
# !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

# Basic dependencies that should work with Python 3.13+
pyttsx3==2.90
psutil==5.9.5
pyyaml>=5.1
tqdm>=4.27.0
numpy>=1.17.0
regex!=2019.12.17

# The following packages are commented out because they are not compatible with Python 3.13+
# Uncomment them if you're using Python 3.10
# --extra-index-url https://download.pytorch.org/whl/cu118
# --find-links https://huggingface.github.io/tokenizers/wheels/tokenizers-0.13.3-cp310-cp310-win_amd64.whl
# tokenizers==0.13.3 --only-binary=:all:
# transformers==4.30.2
# torch==2.6.0
# torchvision==0.17.0
# torchaudio==2.6.0
# huggingface-hub>=0.14.1
